Moving out BacktestingBroker and build from open orders changes
